Change how the token is returned
By adding `.findFirst()` to the stream, and due to the lazy nature of
streams, the `.peek()` method earlier in the Stream's logic will end
as soon as the first match is found. This means the second token will
never be stored/cached. Changed the collection to ALWAYS collect all
elements, and then return the first one. This will ensure that we store
both realm/token combinations for the user.